+++ pleco/benches/magic_benches.rs
