+++ pleco/src/board/castle_rights.rs
