+++ pleco/src/board/piece_locations.rs
