+++ pleco/src/bots/README.md
