+++ pleco/src/core/bit_macros.rs
