+++ pleco/src/core/bit_twiddles.rs
