+++ pleco/src/core/magic_helper.rs
