+++ pleco/src/tools/timer.rs
