+++ pleco/tests/basic_bots.rs
