+++ pleco_engine/src/pleco_searcher/threads.rs
@@ -7,12 +7,12 @@ use std::cmp::{min,max};
 use std::{mem,time};
 use std::time::Duration;
 
-use board::*;
-use core::*;
-use board::eval::*;
-use core::piece_move::BitMove;
-use tools::tt::*;
-use engine::*;
+use pleco::board::*;
+use pleco::core::*;
+use pleco::board::eval::*;
+use pleco::core::piece_move::BitMove;
+use pleco::tools::tt::*;
+use pleco::engine::*;
 
 use super::thread_search::ThreadSearcher;
 use super::misc::*;
